replace global dbs with singleton in integration tests
